Create empty FSA cache clear module
Created the .info and .module files for a new custom module that will
provide clearing functions for Varnish and CloudFlare caches on the Food
Standards Agency website.

The purpose of this module is to overcome the problem where
files that are updated don't get cleared from the Varnish or CloudFlare
caches. With Varnish, this is due to the domain-specific clearing
function built into the contrib Varnish module. Because FSA update
their website via admin.fsa.gov.uk, www.fsa.gov.uk does not get cleared.

[ Partial fix for #2014121910000013 ]