Disable database caching
Under normal circumstances, the FSA website uses memcahce is its cache
backend. However, when using Drush, database caching is erroneously
used instead.

This is the result of checking an Apache environment variable to determine
whether we're on production, staging or dev. Drush doesn't see this
variable, and so the default is used, which is database caching.

Commented out the line in settings.php where this is set.

[ Fixes #2015022410000048 ]