Remove duplicated IDs from FAQ items
Due to a bug in the field_group module, multiple elements
on the same page can be given duplicate IDs, which causes
the page to fail validation.

This behaviour is noted in https://www.drupal.org/node/2037731

Long-term fix is to update the field_group module.

Short-term fix included here is to use an implementation of
hook_field_group_pre_render_alter() to remove the ID attribute
and re-render the wrapper for the field group items.

[ Partial fix for #2015012710000017 ]