Add additional process function for entities
The process function `omega_cleanup_attributes()` causes HTML classes
to be duplicated in the markup for field collection items due to the
markup found in the template for field collection items.

In order to fix this, we add a process function immediately following
`omega_cleanup_attributes()` called `site_frontend_cleanup_attributes()`.

This function checks for duplicated class arrays and removes the one
that's part of the `attributes_array` after making sure we don't
lose any classes that were in one but not the other.

This fixes the duplicated class name problem that was previously in evidence
on FAQ pages, eg http://www.food.gov.uk/science/acrylamide

[ Partial fix for #2015012710000017 ]