Add constant for domain and check base URL
Created a constant `FSA_PASSWORD_UPDATE_BASE_URL` to hold the base
URL and used this later in the callback function to check that we're
on the right domain, as well as for redirecting if the user is already
logged in.

We now check the `$base_url` against the new constant, and if it
doesn't match, we redirect the user. This is a bit belt and braces as
the FSA site already redirects users to admin.food.gov.uk if they try
to get to /admin or /user via www, but it's a nice added bonus.

[ Partial fix for #2015021310000014 ]