Create preprocess function for job nodes
Amended the generic `site_frontend_preprocess_node()` function to call
node type-specific preprocess functions if they exist.

Created a specific preprocess function for job nodes. The main purpose
of this function is to processing on the new URL field that has been
added to Job nodes.

If this new field is populated, the link attached to the title points
to the external URL and opens in a new window. If not, the standard node
URL is used instead.

[ Partial fix for #2015021310000023 ]