Recreate news_section_feature views from prod
The views for the news section are under the control of the feature `news_section_feature`,
which is very out of date with the database configuration on all environments.

Due to a problem with the environment, it's not possible to recreate the feature from
prod. However, since this feature contains only views, the two views in question
were exported from the prod environment and pasted into news_section_feature.views_default.inc

This should enable us to modify the view for allergy alerts and update the feature for deployment
to the dev, staging and production environments.

It's worth noting that the views are not the same on all the environments at the moment, but
taking prod as the master.

[ Partial fix for #2015020510000011 ]