Create custom Zendesk feedback tab module
Added a custom module to replace the standard Zendesk feedback tab
JavaScript with a custom version that resides inside the custom module
folder.

The custom JavaScript file is a copy of the modified version that was
previously added to the contrib module, so includes all the previous
accessibility fixes.

An additional feature is that the module now checks to see whether
we're on an admin path, and if so, does not add the JavaScript, which
means that the feedback tab will not appear in the admin section.

Of course, if we're looking at the admin section via the overlay, then
this does not apply.

[ Partial fix for #2014090110000035 ]