Add new theme function for file icons
In order to fix the problem that file icons are not
getting proper ALT attributes, we override the existing
theme function `theme_file_icon()` with a new function:
`site_frontend_file_icon()`.

This uses a new function `_site_frontend_file_type_names()`
to map MIME types to human-readable names.

The new function contains a default array of mappings, which
can be overridden and augmented via a new theme setting.

[ Partial fix for #2014090110000035 ]