Add ability to specify output format in Drush command
When listing Scottish nodes to be unpublished using the Drush
command `fsa-su`, it is now possible to specify CSV output
format as opposed to Drush table.

The output of the command can then be directed to a file instead
of stdout if required, eg

`drush fsa-su --list --format=csv > /tmp/scottish-pages.csv`

This is useful for identifying which nodes will be unpublished
before running the unpublish command: `drush fsa-su`.

[ Partial fix for #2015030410000011 ]