Add admin field for feedback email subject
Using the admin interface, it's now possible for site editors
to change the subject line of the email message that get sent
on feedback submission.

Also added a new helper function `_fsa_feedback_email_subject()`,
which generates the email subject and is used both in the admin
form and in the email itself.

[ Partial fix for #2015040810000011 ]