Recreate content type meta tag
The Omega theme simplifies the content type meta tag for
HTML5, but this causes errors in the W3C validator as the
tag appears fairly low down the page.

However, if we use the full version of the tag, ie:

`<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />`

the errors no longer appear.

We unset the meta element in `hook_html_head_alter()` and then recreate
it using the code from `_drupal_default_html_head()`.

This fixes the error locally.

[ Partial fix for #2014091510000018 ]