Add new nodeblock template
Created a nodeblock template `node--nodeblock.tpl.php` in site_frontend
as a slightly modified version of the default theme implementation that
ships with the nodeblock module.

The modifications are as follows:

- We remove `class="content"` from the content section as Omega subthemes
  add classes to the attributes array. We therefore apply the `content` class
  via a preprocess function: `site_frontend_preprocess_node()`.
- We remove `class="<?php print $classes; ?> clearfix"` from the main
  container `<div>` as this causes attribute duplication in Omega themes.

These are to overcome issues that arise from Omega duplicating the class
attribute and causing validation to fail.

The function `site_frontend_preprocess_node()` has also been modified to add
back classes that were originally applied directly within the template markup.

[ Partial fix for #2014091510000018 ]