Remove data url and MS filter from feature block h2
In _feature_blocks.scss, commented out the data URL background as
it doesn't seem to work and causes validation issues.

Also, commented out the MS-specific filter as this also doesn't
work properly. Gradient filters seem to apply only if the element
has height.

With these removed, the feature block headers on the homepage
appear, whereas they didn't before as the text was effectively
white on white.

[ Partial fix for #2015041610000014 ]