Create feature for external links
Created a new feature: `external_links` that contains the views
used to display external links and sites in the right-hand column
of certain pages on the FSA site.

Currently contains the following views, as exported and imported from
the production FSA website:

* `external_links`
* `external_sites`
* `external_other`

These will be modified in subsequent commits to enable text to be
automatically translated into Welsh on Welsh-language pages.

[ Partial fix for #2015030910000039 ]