Enable disclaimer on external sites to be translated
The disclaimer in the footer of the `external_sites` view previously
contained markup, which meant that it could not easily be translated
into Welsh when displayed on Welsh-language pages.

Separated out the markup from the text into three distinct footer
elements, and now translation works as expected.

Features to be reverted
-----------------------
* external_sites : `drush fr external_links --y`

Manual configuration required
-----------------------------
* Enable new feature module: external_links : `drush en external_links --y`
* Upload/add appropriate Welsh-language translations.
* Clear all the caches: `drush cc all`

[ Partial fix for #2015030910000039 ]