Replicate styles for alert items
In pages such as /enforcement/alerts, the markup generated by the Views
fields is invalid due to <article> elements appearing within <span>
elements.

The solution is to untick `Provide default field wrapper elements`
in the Field settings for the view. This removes the surrounding
<span> elements.

However, with this markup removed, we need alternative ways of
targetting these elements, so additional styles have been added to
_search_results.scss to replicate the style of the existing page.

This would benefit from improvements as we work through the pages,
and may need to be refactored to account for different types of
search result.

Manual configuration required
-----------------------------
1. Log into Drupal
2. Go to Admin > Structure > Views
3. Search for `Alerts`
4. Click the Edit button
5. Under Format, click the Settings link next to Show: Fields
6. Untick `Provide default field wrapper elements`
7. Click Apply (all displays)
8. Click the Save button
9. Clear the css-js cache: `drush cc css-js`

[ Partial fix for #2015042410000017 ]