Adjust styles for news centre listing
The Page display of the News Centre view was creating
HTML validation errors due to the presence of <article>
elements with <span> elements.

Once the <span> elements are removed by unticking the
`Provide default field wrapper elements` box for this display,
some CSS work is required to ensure that the styling of the
front-end interface remains unaffected.

This builds on the previous work for food alerts, and these
have been checked following this work.

Manual configuration required
-----------------------------
1. Log into Drupal as user 1 (Tharg)
2. Go to Admin > Structure > Views
3. Search for News Centre
4. Click the Edit button
5. Make sure you're on the display Page
6. Under Format, click Settings next to Show: Fields
7. Select This page (override) from the dropdown next to For
8. Untick Provide default field wrapper elements
9. Click Apply (this display)
10. Click save

The manual configuration should be carried out prior to the
code release.

[ Partial fix for #2015042410000017 ]