Modify the way job URLs are constructed
Currently, when extenal URLs are added to job nodes,
they are listed on the jobs page with URLs pointing
to the external site.

Due to the way the links are being rendered, if the
external URL contains a query string, it produces
invalid HTML.

Modified the way the URL is created to use a render array
with theme 'link'.

Tested with various types of link and none. Code now validates.

Modified both the preprocess function and the template
for job nodes.

[ Partial fix for #2015042410000017 ]