Force feedback email links to use HTTPS
At the bottom of the email that site administrators receive when a user
submits feedback, there are two links to the Drupal admin interface.

Forced these to use HTTPS so that site admins can click on them and
go straight into the admin interface without having to authenticate again
if they're already logged in.

[ Partial fix for #2015040810000011 ]