Enable sorting of feedback messages
The form provided by the contrib Feedback module to enable
feedback messages to be viewed was not allowing the columns
to be sorted.

This was because the `orderByHeader()` method of the query object
was not being used.

In order to fix this, it was necessary to rewrite the funcdtion and
incorporate it into the custom module.

[ Partial fix for #2015040810000011 ]