Remove font scale links from accessibility template
Because the text size buttons work only when JavaScript is present
and activated, there's little point in displaying the links if this
is not the case.

Consequently, instead of coding the HTML for the links in the template,
we'll add these in via JavaScript.

This is progressive enhancement in that it adds functionality where
the user agent is capable of making use of it. If we were to leave
the links in, it would get confusing for any users who don't have
JavaSript operational.

[ Partial fix for #1030 ]