Set number of font size controls in module
In the accessibility_links module, we define a constant that
determines the number of font size control buttons that appear
after the link to the accessibility page.

The value of this constant is output as a Drupal JavaScript setting
and is used by the JavaScript function to determine how many links
to create.

[ Partial fix for #1030 ]