Rearrange elements on consultations pages
In order to change the order of elements on consultation pages
to match the request from FSA, it has been necessary to move some
elements programmatically. This is because we need to place some
items provided by context between items provided by the node content.

We include both a Welsh and an English version of the section to
be relocated.

[ Partial fix for #2015020910000031 ]