Ensure language of feedback thankyou message
Because the feedback form uses an AJAX callback, it was not
retaining the language of the calling page, so when submitting
feedback from a Welsh-language page, the thankyou message was
being returned in English.

To get around this problem, we get the value of the thankyou
message on the page where the form appears and store it as a
value field. We then use the value of this field in the custom
submit function to set a message.

[ Partial fix for #2015040810000011 ]