Create preprocess function for views-fields
Created a new preprocess function:

`site_frontend_preprocess_views_view_fields()`

We use this to determine whether any of the items in
the view is a link to the current page, and if so,
we set the `fields` variable to an empty array.

This fixes the issue where the related links
block in the right-hand column of the page
sometimes contained a link back to the page itself.

[ Partial fix for #1046 ]