Change local_get_site_domain() function
Made some changes to the function `local_site_get_domain()`. These
are:

1. Use a separate function `local_get_environment()` to get environment
2. Add an argument `$admin` to specify that the admin version of the domain
   be returned. Defaults to FALSE.

Existing functionalilty should be preserved.

The new functionality will allow us to get an appropriate admin domain
based on the current environment. For example, if we're on staging, the
function will now return '//stagingadmin.food.gov.uk' if we pass TRUE to
the function, but '//staging.food.gov.uk' if we pass FALSE or no arguments.

@todo look at passing the $options parameter to determine protocol. At present
      we are hard-coding protocol-relative URLs, but this doesn't feel as
      flexible as it could be.

[ Partial fix for #1066 ]