Move publication response summary after consultation details
As requested by FSA, moved the Publication of response summary section on
the consultation template to appear after the Consultation details section.

We do this by first determining the weight of the Consultation details group (which is
assigned even if the group is not present), incrementing it and assigning it to the
publication of response summary section.

[ Partial fix for 2015020910000031 ]