Hide consultation details field on node edit form
As requested by FSA, we hide the consultation details field on the
node edit form as this field is no longer required.

For new nodes of this type, we do not show the field. Similarly, for
existing nodes where the field is empty, we hide it. However, for
existing nodes where the field is populated, we do display it to
allow editing.

We achieve this through a new custom module `fsa_consultations`.

New custom modules included
---------------------------
* fsa_consultations

Manual configuration required
-----------------------------
* Enable the fsa_consultations module: `drush en fsa_consultations --y`

[ Partial fix for #2015020910000031 ]