Enable Honeypot on the feedback form
Integrated the Honeypot module with the feedback form.

This was not a straightforward integration due to the custom
form validation in place for the feedback form.

It wwas necessary to implement a custom  function for the
honeypot validation, and so far we have not implemented
the time restriction validation that the module also provides.

This too will require a custom validation function in order to
function correctly with the feedback module. However, it is hoped
that the honeypot validation on its own will suffice.

[ Partial fix for #1076 ]