Add preprocess function for consultation dates field
In order to make the prefix for the consultation dates field
translatable, we create a preprocess function for that field:

`site_frontend_preprocess_field__field_consultation_dates()`

This is called by the main field preprocess function:

`site_frontend_preprocess_field()`

based on the field name.

Currently this affects only the full view mode, as prefix
is not used in any of the other view modes. If we decide in future
to use prefix for the other view modes, we may have to amend this
function.

[ Partial fix for #2015020910000031 ]