Fix markup issue with field_groups module
The field group module introduces some faulty markup in the
function `field_group_pre_render_div()`.

In order to fix this, we implement `hook_field_group_pre_render()`
in the fsa_consultations module. We fix it by correcting the faulty
markup using a simple `str_replace()`.

To make sure our function runs after the pre render function in the
field_group module, we use an implementation of
`hook_module_implements_alter()` to move it to the end of the list.

[ Partial fix for #2015020910000031 ]