Remove styles from maintenance-page.tpl.php
Commented out the $styles variable in the maintenanc-page.tpl.php template
as we're going to be using a single CSS file.

Temporarily hard-coded the reference to maintenance-page.css in the template,
but this will be replaced with a variable and an addition to the preproces
function.

Also created a basic offline maintenance page template that pulls in the static
version of the maintenance page.

This needs more work.