Create new two-column layout with left nav
Created a new two-column layout with left-hand navigation
and wide content column.

It is based very closely on the existing three-column layout.

The only major difference in the template is that the region
sidebar_second is now within the same markup as sidebar_first
so that it appears immediately beneath the navigation in the
left-hand column.

There are also a couple of changes in the SASS/CSS.

[ Partial fix for #10135 ]