Use distinct query for consultation year search
For the custom consultation search, previously the values for
the select box of consultation start years were derived using
a range starting at the earliest year for which we had a published
consultation and ending with the current year.

There were two flaws with this approach:

1. It included Welsh-language consultations, which we didn't want

2. It could potentially include years for which there are no
   published consultations if they are in the middle of the range

Changed the query to select distinct years for which there are
published consultations whose language isn't Welsh.

[ Partial fix for #2014090410000039 ]