Allow manual entry of business
In the report a food problem application, it is
now possible to enter business name and location
manually from the first screen.

Even if a business has been selected via the
autocomplete, if this is replaced with a manual
entry, the selected place is cleared so that it
doesn't affect the subsequent lookup.

The JavaScript is a bit messy, and there is
certainly some redundancy there that could be
removed, but it has been tested across major browsers
and it works as expected.

@todo Clean up the JavaScript and make it more elegant