Include content types in select criteria
When determining whether we have any content items due
for review today, we weren't taking into account the content
types and whether they are currently active. This meant that
an email was being sent even though there were no items to
show since there was an item due for today, but its content
type was disabled for review.

Changed the query to include content type.

[ Partial fix for #10143 ]