Enable admins to edit social media icons
Created a very basic admin interface enabling site administrators
to change the order, number and properties of the social media icons.

The width of the container element is also updated based on the number
of items included.

In addition, it's now possible to specify the icon either by its
URL or by its file ID if uploaded to Drupal manually.

[ Partial fix for 10207 ]