Modify image clearing from CloudFlare
Changed the way images are cleared from the CloudFlare cache.

Instead of responding to the expire event, we now use a custom
submit handler for the file edit form that enables us to determine
which derived image files are in existence, and clear them as well.

This should solve the problem raised in #1069 where image files
were not updating.

[ Partial fix for #1069 ]