Don't exclude current domain from Varnish clearing
Up until now, we've been excluding the current domain from
our Varnish cache clearing function as this is already covered by
the standard Expire and Varnish modules. However, when we're
clearing Varnish using Cron, this means that we miss out the current
domain, which means that we don't regenerate the derived images.

Temporarily just comment it out, but we may decide to include an
argument to specify whether it's excluded or not. In reality, the
performance hit of clearing it twice isn't really a big deal.

[ Partial fix for #1069 ]