Remove $ at end of Varnish purge URLs
This may be causing problems with URLs that contain
query strings, such as derived image URLs with their
itok parameters.

This isn't a permanent solution, even if it works,
because it may cause issues with clearing the cache for
more pages than we'd like. Instead we probably need to
look at either pattern matching query strings in URLs or
including the query string when we add image URLs to the
queue to be cleared.

@todo Find proper way of doing this.

[ Partial fix for #1069 ]