Enable in-page links to the feedback form
Any link to the feedback form using the href `#feedback-form` will
now open up the feedback form and scroll smoothly down the page to
its location.

If JavaScript is not enabled, the form will already be displayed,
and the link will just go to that bookmark, thereby achieving the
same result.

@todo Include the page title in the feedback emails and reports

[ Partial fix for #10229 ]