Upgrade Migrate module to 7.x-2.8
Upgraded the Migrate module from 7.x-2.5 to 7.x-2.8.

This is a security release.

As this module was used during migration and not during the
operation of the site, there should be no real need for
significant testing.

However, some things have changed in the module's API, so if
there is ever a need to use it again, it is worth checking in
CHANGELOG.txt for required changes.

[ Fixes #10253 ]