Upgrade Inline Entity Form to 7.x-1.6
Upgraded the Inline Entity Form module to version 7.x-1.6.

In addition to the security fix, there is also a fix for a
bug whereby if multiple entity edit forms are open in a given
edit form, clicking save on one, closes them all, which means
that any saved data in the others is lost.

@see https://www.drupal.org/node/2343689

Full release notes here:

https://www.drupal.org/node/2507593

Entity reference widgets are used on the following content
types:

* General page (document_page)
* Contact group (contact_group)
* Contact list (contact_list)

They are also in use on the FAQ field collection: field_fc_qanda

[ Fixes #10111 ]

This is a security fix.