Check review date format before converting
When saving the review date, we were converting it to a timestamp,
assuming that the input was a value from the form, which is a string.
However, in bulk operations where a node already has a review date set,
the review date will already be a timestamp, so trying to convert it
causes an error.

To fix it, we check first whether the value is numeric, and if so,
we use it as is.

[ Fixes #10262 ]