Upgrade Drupal Core to version 7.41
Upgraded Drupal Core to the latest version: 7.41.

This is a security release.

Database updates required
--------------------------
A number of database updates are required as part of
this upgrade. These may be applied as part of the
deployment process, but if not, following successful
deployment, log into the Drupal server via SSH, navigate
to the Apache root folder and run:

`drush updb --y`

This will apply any remaining updates, as well as clearing
the Drupal caches.

[ Fixes #10269 ]