Add file upload help text
Added the ability for filename validators to provide help
text for display on file upload forms.

This is done by overriding the existing theme function for
file upload help text and then adding elements to each of
the validator definitions.

Help text can be provided as a string or by a callback function
in a similar way to error messages.

[ Partial fix for #10282 ]