Allow feedback messages to capture query strings
Currently, the feedback module captures only the path of a page, not
the query string. However, it is useful to be able to see the query
string for failed searches and similar pages.

Modified the form to capture the query string, and also used
hook_feedback_insert() to modify the url property of the feedback
entry - if necessary.

[ Partial fix for #10287 ]