Change implementation for SiteImprove meta tags
Was previously using hook_node_view() to add the meta tags
to the page. However, on the FSA site, this was being triggered
by an embedded node that was being displayed in 'full' view mode.

Instead, use an implementation of `template_preprocess_page()` to
ensure that the node is the main page one.

[ Partial fix for #10289 ]