Remove dependency on specific Redirect module version
Previously, the National Archives module required the most recent
version of the Redirect module as it made use of the status property
of redirect entities, which was not available before version
7.x-1.0-rc2.

Now, we check to see whether the function `redirect_disable_by_path()`
exists, and if so, we use it to disable the redirect. Otherwise, we
simply delete it.

Also cleaned up some white space and improved commenting.

[ Partial fix for #2014090110000295 ]