Add archiving for files
When files are deleted, a redirect to the National Archives is automatically
created.

Similarly, when a file is created, Drupal will check to see if a redirect exists
with the same internal URL, and if so, the redirect will be deleted.

This latter step is actually somewhat redundant as a file that exists will always
take precedence over a redirect thanks to Drupal's .htaccess rewrite rules. However,
it's neater that way.

[ Partial fix for #2014090110000295 ]