Upgraded Views Bulk Operations to 7.x-3.3
Upgraded the Views Bulk Operations module to version 7.x-3.3.

This is a security release.

Database schema updates
-----------------------
There are no database schema updates for this version.

Testing
-------
The functionality provided by this module is used primarily in the
content administration screen in the Drupal admin interface:

https://admin.food.gov.uk/admin/content

It provides fnctionality via the bulk update dropdown.

This should be tested once the module has been upgraded.

[ Partial fix for #10297 ]