Add archived file interstitial page
When a file has been archived, we display a page indicating
this and providing a link to the file on the NA site.

If the file does not exist on the NA site, we display a
404 page instead.

Also added a helper function to determine whether a path is
for a file.

[ Partial fix for #2014090110000295 ]