Don't show feedback form for specified paths
In response to the request to prevent feedback emails
from spoof registration pages, added the ability to specify
an array of paths for which the feedback form will not be
displayed.

At present, this array is limited to pages with paths of the
form ^.*/user/register$. This will exclude paths such as:

/news-updates/news/2011/4574/user/register

which is one of the main culprits.

The array is currently hard-coded in the FSA Feedback module,
but if necessary, a UI could be provided to allow administrators
to designate other paths - though this may require regex knowledge.

Alternatively, the feedback form could be inserted using Context,
by exposing it as a block. This would enable administrators to
exclude it via the existing Context UI. For now though, this should
suffice.

[ Partial fix for #10303 ]