Don't use check_plain() for business names
When creating the aria-lable and title attributes for submit
buttons on the business list form, don't prefix the placeholders
with @ as this causes the t() function to run them through
check_plain(), which results in apostrophes being converted to
HTML entities. This sounds weird when read by a screen reader.

Instead, use htmlspecialchars() directly, specifying ENT_COMPAT
instead of ENT_QUOTES.

[ Partial fix for #10273 ]