Remove view-dom-id class for node header fields
Used an implementation of template_preprocess_views_view() to remove
the `view-dom-id...` class from the node header fields view, which
appears at the top of many nodes.

This class is used primarily with exposed filters, and is consequently
not used on this view.

Because the class name changes frequently, it causes false positives
for the GovDelivery HTML PageWatch mechanism.

No other views are affected.

[ Partial fix for #10225 ]