Change how we interpret URL patterns
Previously, for URL paths that don't correspond to an
actual alias, we were just looking at the first part of the
path to determine what action to take and which variables to
set. We now look at the start of the path and do the same. This
will help when determining local authority URLs.

@todo Determine whether this will be necessary, depending on CTools object cache

[ Partial fix for #10273 ]