Get local authority ID from query string
On the final stage of the local authority search, if we don't have
a local authority by way of the form state, we now look in the query
string to see if one has been set via hook_init(). If so, we use that.

We also test using CTools object cache to store the local authority, though
this isn't yet used on the page.

@todo Replace query string implementation with CTools object cache

[ Partial fix for #10273 ]