Upgrade context to version 7.x-3.6
Upgraded the context module to version 7.x-3.6.

Patch status
------------
This module had previously been patched to improve performance.
The patch may have found its way into the main codebase, but if not,
and if performance declines, then we may need to re-apply the patch.

Testing
-------
The context module provides much of the block postitioning functionality
on the FSA site, so this needs thorough test for each defined context.

This should include testing the editing functionality via Context UI.

[ Fixes #1086 ]