Enable forced checking of APIs
Typically, the external APIs are checked and then assumed to be in
their current state (healthy or unhealthy) for specified periods of
time to avoid over-checking.

Added a parameter to the checking function to enable forced checking,
even if we're within the healthy or unhealthy time window.

Also get list of APIs from helper function.

[ Partial fix for #10273 ]