Modify MapIt authority types function
This helper function used to return just the selected authority types,
which were set in the code. Now, it returns a default set unless a variable
is set, in which case it returns that.

A new parameter `$full` tells the function to return all authority types as
an associative array for use with checkboxes that can determine which types
are searched.

[ Partial fix for #10273 ]