Use locale() to make strings translatable
In order for certain strings of text to appear in Welsh language,
we use the `locale()` function to output the text.

Note that we do not use `t()` as these are not static strings, but
variables coming from other sources beyond our control.

Note also that the translations for these strings will need to be
imported since Drupal will not pick them up automatically for the
most part.

[ Partial fix for #10273 ]