Upgrade Display Suite to version 7.x-2.12
Upgraded the Display Suite contributed module to version 7.x-2.12

This is a security release.

Database schema updates
-----------------------
The following updates are required as part of this upgrade:

* 7202 -   Increase the label storage length to 128.
* 7203 -   Increase the view_mode storage length to 64

Be sure to run `drush updb --y` as part of the deployment process.

Testing
-------
Display suite is used across the site to provide layouts and rendering
of pages. I have tested this upgrade on my local install and have been
through the code to review the changes. No problems were found, and the
majority of the code changes were minor and unlikely to cause problems.

When testing this module upgrade, it is important to view pages of
various different content types to ensure that the layout appears
unaffected.

[ Fixes #1082 ]