Replace views exposed filter on search results page
We replace the views exposed filter keywords field with a new form
builder form as the exposed filter doesn't play nicely with the
Solr spell checking.

However, at present, this means that the spell checking module must
be enabled in order for this to work, so we need to put in some logic
such that if the spell checker module isn't enabled, then the old
version is used.

@todo Change logic so old version used if spell checker not available

[ Partial fix for #10176 ]