Allow carousel text box to expand to fit text
Previously the width of the text box on the homepage carousel was
fixed at 400px. Removed this directive to enable the box to expand
with the text, as requested by Sally.

Tested on Firefox, Safari, Opera and Chrome on OSX. Also tested on
IE9, Firefox and Chrome on Windows 7. All working correctly.

Carousel images are not displaying properly on IE8, but this is
the case for the existing site as well.

@todo Look at size of text box for IE8

[ Partial fix for #10345 ]