Remove insert media button from CKEditor toolbar
Removed the insert media button from the CKEditor toolbar as it
is not used by FSA and no longer works properly with the upgraded
version of the CKEditor module (7.x-1.17).

Feature to be reverted
----------------------
In order for the change to take effect, the feature `ckeditor___editor`
will need to be reverted. The Drush command is below:

`drush fr ckeditor___editor --y`

After this, it's worth clearing the caches once more:

`drush cc all`

[ Partial fix for #1084 ]