Add replacement callback function for rating AHAH call
The AHAH callback provided by the Rate module does not directly
support multiple languages. This means that when it returns the
current rating, it is always in English.

Provided an alternative callback function that takes language code
as a parameter and then sets the $language global variable before
calling the original `rate_vote_ahah()` function.

[ Partial fix for #10346 ]