Add access callback for food problem reports
In order to be able to view food problem reports, users currently
need access to the Drupal admin Reports menu. This may not be
desirable for all users.

To fix this, we are going to add a new menu item to the main admin
menu to provide access to food problem reports for those users with
a special role.

But we don't want all admin users to see this item if they already
have access to the reporting section, so a new access callback is
being added to check this first.

[ Partial fix for #10393 ]