Add implementation of hook_views_post_execute()
In order to modify the HTTP headers associated with the exported
CSV data for the report a food problem module, we implement
`hook_views_post_execute()` and modify this setting. This changes
the filename associated with the downloaded file, as well as adding
a content-disposition header, which may help Excel on Windows to
recognise this as a CSV file.

[ Partial fix for #10405 ]