Modify option labels in problem reports view
The option labels for the views exposed filters for manual
submission and email sent are set to 'True' and 'False' by default.

To tie in better with the values used in the table, we want them
to be 'Yes' and 'No' respectively.

As this isn't provided by the views filter handler, rather than
creating a new handler, we use an implementation of
`hook_form_FORM_ID_alter()` to replace the filter options instead.

[ Partial fix for #10405 ]