Create alternative token replacement function
In an attempt to circumvent any potential caching issues with the
FSA GUID tokens that may have caused the tokens not to be replaced
in the RSS feeds, we create a new function to be used in place of
Drupal's built-in `token_replace()` function for this module.

This function is much simpler and does not employ any caching.

[ Partial fix for #10866 ]