Allow unpublished nodes in ref fields. 102052
By default, Entity reference fields will not allow unpublished nodes
to be included. If an unpublished node is added, a validation handler
will prevent the node edit form from submitting.

To get around this, we implement a query tag alter hook and remove
the condition that requires the node to be published.

[ Partial fix for #102052 ]