Catch exceptions getting entity info. 102408.
We need to be careful when trying to get information about objects
that are being expired. Sometimes - as in the case of menu lnks -
they are not entities or even objects at all.

Rather than try to work out whether an object is an entity, we
just wrap that call to entity_extract_ids() in a try... catch
block.

[ Partial fix for #102408 ]